refactor: printing
