+++ packages/web/src/features/printing-business/frames/PrintingDetailFrame.tsx
@@ -7,7 +7,7 @@ import {
 } from "@sparcs-clubs/web/common/components/ListItem";
 import Typography from "@sparcs-clubs/web/common/components/Typography";
 
-const ManagePrintingDetailFrame = () => (
+const PrintingDetailFrame = () => (
   <FlexWrapper direction="column" gap={20}>
     {/* TODO: 아래 정보들 백 연결하기 */}
     <FlexWrapper direction="column" gap={16}>
@@ -51,4 +51,4 @@ const ManagePrintingDetailFrame = () => (
     </FlexWrapper>
   </FlexWrapper>
 );
-export default ManagePrintingDetailFrame;
+export default PrintingDetailFrame;
